Revert "Update dependencies for 1.7.10"
This reverts commit 420b8b4bcc32d163f2ed6025c391f2d5ff7bc158.

+++ dependencies/main/java/ic2/api/energy/tile/IEnergyConductor.java
@@ -19,7 +19,7 @@ public interface IEnergyConductor extends IEnergyAcceptor, IEnergyEmitter {
 	 * 
 	 * @return Insulation energy absorption in EU
 	 */
-	double getInsulationEnergyAbsorption();
+	int getInsulationEnergyAbsorption();
 
 	/**
 	 * Amount of energy the insulation will handle before it is destroyed.
@@ -27,14 +27,14 @@ public interface IEnergyConductor extends IEnergyAcceptor, IEnergyEmitter {
 	 *
 	 * @return Insulation-destroying energy in EU
 	 */
-	double getInsulationBreakdownEnergy();
+	int getInsulationBreakdownEnergy();
 
 	/**
 	 * Amount of energy the conductor will handle before it melts.
 	 * 
 	 * @return Conductor-destroying energy in EU
 	 */
-	double getConductorBreakdownEnergy();
+	int getConductorBreakdownEnergy();
 
 	/**
 	 * Remove the conductor's insulation if the insulation breakdown energy was exceeded.
