+++ dependencies/main/java/ic2/api/info/IEnergyValueProvider.java
@@ -7,11 +7,8 @@ public interface IEnergyValueProvider {
 	 * Determine the energy value for a single item in the supplied stack.
 	 * The value is used by most machines in the discharge slot.
 	 * 
-	 * This only applies to basic single use items, others are to be queried
-	 * through e.g. ElectricItem.manager.getCharge().
-	 * 
 	 * @param itemStack ItemStack containing the item to evaluate.
 	 * @return energy in EU
 	 */
-	double getEnergyValue(ItemStack itemStack);
+	int getEnergyValue(ItemStack itemStack);
 }
